backend/requirements.txt
absl-py
asgiref
cachetools
certifi
chardet
charset-normalizer
click
colorama
contourpy
cramjam
cycler
Django
django-cors-headers
django-environ
django-filter
django-rest-framework-mongoengine
django-versatileimagefield
djangorestframework
djangorestframework-simplejwt
djongo
dm-tree
dnspython
environ
fonttools
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
googletrans
h11
h2
h5py
hpack
hstspreload
httpcore
httplib2
httpx
hyperframe
idna
joblib
keras
kiwisolver
libmagic
markdown-it-py
matplotlib
mdurl
ml-dtypes
mongoengine
namex
nltk
numpy
packaging
pandas
pillow
protobuf
pyasn1
pyasn1-modules
Pygments
PyJWT
pymongo
pyparsing
python-dateutil
python-magic
python-magic-bin
python-snappy
pytz
regex
requests
rfc3986
rich
rsa
scikit-learn
scipy
setuptools
six
sniffio
SpeechRecognition
sqlparse
threadpoolctl
tqdm
typing_extensions
tzdata
uritemplate
urllib3
winkerberos
